requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
asyncpg==0.28.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
